# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "效率仪表盘后端API，提供任务、提交、用户、仓库、项目、组织等数据管理接口",
        "title": "Efficiency Dashboard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9990",
    "basePath": "/api",
    "paths": {
...
        "/analysis/git": {
            "get": {
                "description": "根据仓库ID查询Git分析结果，包含提交统计和AI估算信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git"
                ],
                "summary": "获取Git分析结果",
                "parameters": [
                    {
                        "type": "string",
                        "description": "仓库ID",
                        "name": "repo_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "开始日期(YYYYMMDD)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期(YYYYMMDD)",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.GitAnalysisResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
        "main.EstimationInfo": {
            "type": "object",
            "properties": {
                "final": {
                    "type": "number"
                },
                "from_git": {
                    "type": "number"
                },
                "from_task": {
                    "type": "number"
                }
            }
        },
...
        "main.GitAnalysisResponse": {
            "type": "object",
            "properties": {
                "analysis_date": {
                    "type": "string"
                },
                "estimation": {
                    "$ref": "#/definitions/main.EstimationInfo"
                },
                "git_analysis_file": {
                    "type": "string"
                },
                "git_stats": {
                    "$ref": "#/definitions/main.GitStatsInfo"
                },
                "repo_id": {
                    "type": "string"
                }
            }
        },
        "main.GitStatsInfo": {
            "type": "object",
            "properties": {
                "commit_count": {
                    "type": "integer"
                },
                "contributor_count": {
                    "type": "integer"
                },
                "files_changed": {
                    "type": "integer"
                },
                "lines_added": {
                    "type": "integer"
                },
                "lines_deleted": {
                    "type": "integer"
                }
            }
        },
...
    }
}